requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiopath==0.6.11
discord==2.3.1
Pillow==10.0.0